refactor: split log page into row, jump band and empty state
The file was over the length guideline before these changes and the added
animation state pushed it further.
